+++ packages/x_storage_core/lib/x_storage_core.dart
@@ -4,3 +4,6 @@ export 'src/x_storage.dart';
 export 'src/x_storage_provider.dart';
 export 'src/x_uri.dart';
 export 'src/x_storage_type.dart';
+export 'src/x_storage_exception.dart';
+
+export 'src/providers/asset_storage_provider.dart';