+++ packages/x_storage_presigned_url/lib/src/presigned_url_storage_exception.dart
@@ -1,0 +1,6 @@
+import 'package:x_storage_core/x_storage_core.dart';
+
+class HttpException extends XStorageException {
+  final String message;
+  HttpException(this.message);
+}