+++ packages/x_storage_presigned_url/lib/x_storage_presigned_url.dart
@@ -1,3 +1,4 @@
 library x_storage_presigned_url;
 
 export 'src/presigned_url_storage_provider.dart';
+export 'src/presigned_url_storage_exception.dart';